refactor(arch): extract file discovery from IncludeGenerator to data layer
Move file I/O operations from output layer to data layer for proper
separation of concerns per 3-layer architecture:

- Create CnxFileResolver in data/ with static methods:
  - findCnxFile() - search for .cnx files in paths
  - getRelativePathFromInputs() - calculate relative paths
  - cnxFileExists() - check file existence

- Update IncludeGenerator to delegate to CnxFileResolver
- Remove fs import from IncludeGenerator (now pure generation)
- Bump version to 0.1.52
